tests/requirements.txt
pytest-asyncio==0.12.0
aiohttp==3.7.2
pythonping==1.1.1
httpx==0.23.0
aioredis==1.3.1
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.38.3
asyncpg==0.26.0
psycopg2-binary==2.9.3
databases[postgresql]==0.6.1
alembic==1.12.0
pydantic==1.9.0
python-dotenv==0.20.0